Correction précision minicarte (espace vide droite/bas) Daims broutent et lents

+++ 2022_RTS/RTS_client_B41/RTS_vue.py
@@ -431,7 +431,7 @@ class Vue():
                                                   #tags=("mobile","",i.id,)
 
         self.modele.listebiotopes=[]
-        minitaillecase=int(self.tailleminicarte/self.modele.taillecarte)
+        minitaillecase=self.tailleminicarte/self.modele.taillecarte
         couleurs={0:"",
                   "arbre":"light green",
                   "eau":"light blue",
